pyproject.toml
[tool.poetry]
name = "django_backend_starter_template_project"
version = "1.0.0"
description = "django backend starter template project"
authors = ["SoungRyoul Kim <KimSoungRyoul@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.11.0"

# django
django = "^4.2"
django-filter = "^2.4.0"
django-extensions = "^3.0.9"
channels = "^3.0.3"

# django rest framework
djangorestframework = "3.14.0"
drf-spectacular = "^0.26.2"
drf-nested-routers = "^0.92.1"
djangorestframework-simplejwt = "^5.1.0"

# type
mypy = "^1.2"
django-stubs = { extras = ["compatible-mypy"], version = "^4.2.0" }
django-stubs-ext = "^4.2.1"
djangorestframework-stubs = "^1.4.0"
django-ninja = "^1.2.2"



[tool.poetry.group.prod.dependencies]
uvicorn = "^0.17.5"
gunicorn = "^20.1.0"


[tool.poetry.group.mysql.dependencies]
#mysqlclient = "^2.1.1"
pymysql = "^1.0.3"
types-pymysql = "^1.0.19.7"
django-mysql = "^3.12.0"



[tool.poetry.group.postgresql.dependencies]
psycopg = { extras = ["binary", "pool"], version = "^3.1.8" }
django-postgres-extra = "^2.0.8"


[tool.poetry.group.file-upload.dependencies]
django-storages = "^1.13.2"
pillow = "^9.5.0"


[tool.poetry.group.redis.dependencies]
django-redis = "^5.2.0"
redis = "^4.5.5"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.9.3"
Werkzeug = "^2.0.3"


[tool.poetry.group.test.dependencies]
freezegun = "^1.2.2"


[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'


[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]


[tool.django-stubs]
django_settings_module = "config.settings"

